app/requirements.txt
flask
pandas
geopandas
json
requests
matplotlib.pyplot
io
base64